Refactor: Moving all the components under app.js to be under material.js
